Moved openForm(),openSavedForm and prefillAndOpenForm() to FormsInteractor (#54)
* Moved openForm(), openSavedForm() and prefillAndOpenForm() to FormsInteractor.
