+++ odk/extension/src/main/java/io/samagra/odk/collect/extension/interactors/FormsInteractor.kt
@@ -48,4 +48,15 @@ interface FormsInteractor {
      * Note: This modifies the original form described by the form path.
      */
     fun updateForm(formPath: String, values: HashMap<String, String>,listener: FormsProcessListener?)
+
+    /** Opens the latest version related to the formId. Deletes any
+     *  saved instance of a form with this particular formId. */
+    fun openForm(formId: String, context: Context)
+
+    /** Opens a saved form. If no saved instance is found, opens a new form. */
+    fun openSavedForm(formId: String, context: Context)
+
+    /** This method pre-fills a form with data from a given map of key-value pairs,
+     * and then opens the form in a given Android context. */
+    fun prefillAndOpenForm(formId: String, tagValueMap: HashMap<String, String>, context: Context)
 }
