+++ odk/extension/src/main/java/io/samagra/odk/collect/extension/interactors/ODKInteractor.kt
@@ -25,12 +25,4 @@ interface ODKInteractor {
     /** Resets everything in odk and deletes all data. */
     fun resetODK(listener: ODKProcessListener)
 
-    /** Opens the latest version related to the formId. Deletes any
-     *  saved instance of a form with this particular formId. */
-    fun openForm(formId: String, context: Context)
-
-    /** Opens a saved form. If no saved instance is found, opens a new form. */
-    fun openSavedForm(formId: String, context: Context)
-
-    fun prefillAndOpenForm(formId: String, tagValueMap: HashMap<String, String>, context: Context)
 }
