+++ sample/src/main/java/io/samagra/oce_sample/ODKFeatureTesterActivity.kt
@@ -12,6 +12,7 @@ import android.widget.Toast
 import io.reactivex.rxjava3.disposables.CompositeDisposable
 
 import io.samagra.odk.collect.extension.interactors.FormsDatabaseInteractor
+import io.samagra.odk.collect.extension.interactors.FormsInteractor
 import io.samagra.odk.collect.extension.interactors.FormsNetworkInteractor
 import io.samagra.odk.collect.extension.interactors.ODKInteractor
 import io.samagra.odk.collect.extension.listeners.FileDownloadListener
@@ -40,6 +41,7 @@ class ODKFeatureTesterActivity : AppCompatActivity(), View.OnClickListener {
     private lateinit var progressBar: ProgressBar
 
     private lateinit var odkInteractor: ODKInteractor
+    private lateinit var formsInteractor: FormsInteractor
     private lateinit var networkInteractor: FormsNetworkInteractor
     private lateinit var formsDatabaseInteractor: FormsDatabaseInteractor
 
@@ -78,6 +80,7 @@ class ODKFeatureTesterActivity : AppCompatActivity(), View.OnClickListener {
                 currentProjectProvider.getCurrentProject().name
                 formsDatabaseInteractor = ODKProvider.getFormsDatabaseInteractor()
                 networkInteractor = ODKProvider.getFormsNetworkInteractor()
+                formsInteractor = ODKProvider.getFormsInteractor()
                 downloadFormsButton.isEnabled=true
                 downloadAllFormsButton.isEnabled=true
                 clearAllFormsButton.isEnabled=true
@@ -124,7 +127,7 @@ class ODKFeatureTesterActivity : AppCompatActivity(), View.OnClickListener {
                 val formId: String = openFormsInput.text.toString().trim()
                 if (formId.isNotBlank()) {
                     progressBar.visibility = View.VISIBLE
-                    odkInteractor.openForm(formId, context)
+                    formsInteractor.openForm(formId, context)
                 }
             }
             R.id.download_form_button -> {
@@ -190,7 +193,7 @@ class ODKFeatureTesterActivity : AppCompatActivity(), View.OnClickListener {
                 val formId: String = openSavedInput.text.toString().trim()
                 if (formId.isNotBlank()) {
                     progressBar.visibility = View.VISIBLE
-                    odkInteractor.openSavedForm(formId, context)
+                    formsInteractor.openSavedForm(formId, context)
                 }
             }
         }
